Initial implementation of the Face Recognition app integration
At this point, we completely reuse the Recognize view, and for now,
Recognize has priority. If it is installed, it shows the results of
this application.

So if recognize is disabled and Face Recognition is enabled, you can
see the tagged people on that app, you can see the photos of these
people, and the rectangles of the faces.

Still can't add names, hide people, or merge clusters with the current
view since it is intended for the recognize DAV api.

We probably need to add a custom view for this app. Meanwhile, the
side panel that works in both Files and Photos also works in Memories.
You can add the names there, hide the people, and it should probably
only be done there.