Fix FaceRecognition integration doesn't work with postgresql database.
Following the discussion in #304, I tried this solution using the
strict mode of mariadb/mysql.

Split the Clusters and People queries to minimize the considerations
of both functions. Otherwise I had to add many nested ifs.

Also change to only show 15 clusters (as a small optimization) since
there is no point in showing everything, and it is better to show these
little by little.