Implement merge person/clusters on Face recognition backend. Part of issue #285
Just try to make the patch as simple as possible without changing a single line
of behavior in recognize.

The warning that you can only merge with named people is important, and while I
personally don't understand why we would merge unnamed people, it is a feature
that the recognize backend supports.